LayoutTests/media/media-session-restrictions-expected.txt

mediaSessionRestrictions["video"] = ""
mediaSessionRestrictions["audio"] = ""
mediaSessionRestrictions["videoaudio"] = ""
mediaSessionRestrictions["webaudio"] = ""
END OF TEST
